requirements.txt
--find-links https://download.pytorch.org/whl/torch_stable.html
numpy>1.20
torch==2.0.0+cu118
git+https://github.com/minerllabs/minerl
imitation==0.3.2
coloredlogs
matplotlib
gym3
attrs
opencv-python
videoio
